Rewritten SA pools sources. Highly improved performance and coding style. Also extended the CPools interface to support more interface functions and removed the references to game_sa from it. Fixed a switch-case statement not breaking in fire code. It might be the problem that caused freezes.

+++ trunk/MTA10/sdk/game/CEntity.h
@@ -105,7 +105,7 @@ public:
     virtual void                        SetStaticWaitingForCollision ( bool bStatic ) = 0;
 
     virtual void                        GetImmunities ( bool & bNoClip, bool & bFrozen, bool & bBulletProof, bool & bFlameProof, bool & bUnk, bool & bUnk2, bool & bCollisionProof, bool & bExplosionProof ) = 0;
-    virtual int                         GetArrayID ( void ) = 0;
+    virtual unsigned long               GetArrayID ( void ) = 0;
 };
 
 #endif
